clean up

+++ db/schema.sql
@@ -12,7 +12,7 @@ DROP TABLE IF EXISTS role;
 CREATE TABLE role(
 	id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
 	title varchar(30) NOT NULL,
-    salary INT NOT NULL,
+    salary DECIMAL NOT NULL,
     department_id INT NOT NULL,
     FOREIGN KEY (department_id)
     REFERENCES department(id)
